requirements.txt
pyyaml==5.1
markdown==3.3.3
pymdown-extensions==8.1
